Add initial prototype for virtual hardware plugin
This plugin (as with the others) borrows heavily from existing
projects. Minor doc updates have been applied to reflect this
new plugin, but additional work is needed.

This plugin does not currently accept threshold flags to
specify conditions for `WARNING` or `CRITICAL` states. See
GH-33 for future work to add that support. Instead, this
version of the plugin assumes any deviance from the latest
(detected) version is a `WARNING` state.

refs GH-15