Initial draft of host/datastore/vms pairing plugin
This is a functional plugin responsible for verifying
that each VM is housed on a datastore (best) intended
for the host associated with the VM.

By default, the evaluation is limited to powered on VMs,
but this can be toggled to also include powered off VMs.

The association between datastores and hosts is determined
by a user-provided Custom Attribute. Flags for this plugin
allow specifying separate Custom Attribute names for
hosts and datastores along with optional separate prefixes
for the provided Custom Attributes.

This allows for example, hosts to use a `Location` Custom
Attribute that shares a datacenter name with datastores
using the same `Location` Custom Attribute. If not
specifying a prefix separator, the plugin assumes that
a literal, case-insensitive match of the `Location` field
is required. If a prefix seperator is provided, then the
separator is used to retrieve the common prefix for the
`Location` Custom Attribute for both hosts and datastores.

This is intended to work around hosts that may include both
the datacenter name and rack location details in their
Custom Attribute (e.g., `Location`).

This plugin optionally allows ignoring a list of datastores,
and both hosts and datastores that are missing the specified
Custom Attribute.

In addition to specifying separate Custom Attribute names
(required) and prefix separators (optional), the plugin also
accepts a single Custom Attribute used by both hosts and
datastores and an optional prefix separator, also used by
both hosts and datastores.

If specifying a shared Custom Attribute or prefix,
per-resource Custom Attribute flags are rejected
(error condition).

Known issues:

- While testing indicates that this plugin is functional,
  significant refactoring is still needed

- As with other plugins, documentation is incredibly thin and
  examples are missing from the README

refs GH-8