Remove sessions cache support
This is done in order to remove the attempt by the
`govmomi/sessions/cache` package to create a file cache
within the `$HOME/.govmomi` directory path for the user
(`nagios` in our case) executing plugins from this project.

Based on light reading, the session cache is designed
specifically to act as a file-based cache. We either require the
sysadmin using these plugins to create the `$HOME/.govmomi`
directory and set proper permissions, or we find a way to set an
environment variable for that user to point to a different path.

Either requires more Nagios console setup work than intended,
so for now we do without it and attempt to ensure that we
explicitly logout of the current session when the plugin
execution completes.

refs GH-37
refs https://pkg.go.dev/github.com/vmware/govmomi@v0.24.0/session/cache