Fix project name
Update constants values to use correct project name. Remove
duplicate value from project URL constant and reuse app name
constant. This is done in case I use *this* project as a
template for my next one so that I won't end up changing one
value and forgetting to do the same for the other (as I did
here in this project).