Use type switch var to safely call pc.RetrieveOne
Prior to this commit, the `RetrieveOne` PropertyCollector
method was called with an empty interface. This results in a
panic with "unexpected type" as the error string.

Fix by using type switch variable and moving the method call
into each case statement so that the appropriate type is
passed to the `RetrieveOne` method.