datastore: Replace tabwriter VMs table with list
The use of pre tags and tabwriter table is incompatible
with stock Nagios configuration. To use pre tags effectively,
the `illegal_macro_output_chars` list would need to be
modified to allow banned characters.

Instead, we generate a list of VMs with the same information
as before, but in a condensed format.

refs GH-67