Fix age not yet exceeded snapshots display logic
Remove state checks at the snapshot summary sets level,
rely on individual snapshot state checks to determine
inclusion in the "not yet exceeded" list.

The net result is that the snapshots summary report is
generated as intended. As noted previously, further
refactoring is warranted for a later point in time.

refs GH-81