Include VMs outside visible Resource Pools
Historically, the plugins used to monitor our vSphere
environments excluded the default hidden 'Resources' Resource
Pool. This was done because it threw off calculations due
to inclusion of duplicate VirtualMachine entries.

This change works around that by limiting evaluation of the
'Resources' pool to only when the explicit `include` list is
empty and explicitly de-duping the final list of VirtualMachine
entries.

This change should resolve the unintentional exclusion logic
for multiple plugins:

- `check_vmware_tools`
- `check_vmware_hs2ds2vms`
- `check_vmware_snapshots_age`
- `check_vmware_vhw`
- `check_vmware_vcpus`

refs GH-87