Add initial draft of VM power cycle uptime plugin
In short, this plugin monitors the power cycle uptime for Virtual
Machines. In addition to reporting current power cycle uptime, this
plugin also reports which VMs are on the host (running or not) and the
uptime for each.

Doc updates have been applied, example usage has been added, including
a command definition "contrib" file illustrating how the plugin would
be referenced within a production Nagios configuration.

refs GH-48