Triggered Alarms | Allow 1+ DCs, fail on missing
The initial plugin release was intended to support retrieving
triggered alarms from a user-specified datacenter or if not
specified, from all datacenters. Unfortunately the initial
plugin release did not properly filter triggered alarms to the
user-specified datacenter. Regardless of whether the user
specified a datacenter all datacenters were evaluated.

This commit adds support for specifying one or more datacenters
and filters accordingly. Additionally, this commit *also*
reverses the decision to fallback to the default datacenter
if the specified list is not found and instead considers
any missing requested datacenter names as a CRITICAL error.

Lastly, this commit includes various debug logging tweaks
and minor adjustments to the Long Service Output shown
as "extended output" for the check results. These adjustments
focus on exposing which datacenters were requested, which
were evaluated and which contain triggered alarms.

refs GH-230