Move RPs include/exclude list validation
Move this validation to each plugin-specific validation
block that needs to use it. With upcoming support for
filtering Triggered Alarms based on entity Resource Pools
using similar validation, having this check as a "common"
item doesn't work as well anymore.

While leaving it as-is doesn't break anything, it doesn't
make as much sense while reviewing the code.