Reorder cmd def params, update host svc checks
While adding the missing DCs args to the host service
checks, I opted to reorder the command definition flags
in order to place the include/exclude lists at the
end. This allows the service checks to be defined with
the longer lists of details (the include/exclude lists)
also at the end.

The assumption is that this order will more closely
match existing service checks that have lists of
values at the end.

refs GH-268