Update check_vmware_datastore to emit PerfData
Add initial, prototype support for Performance Data
output to the `check_vmware_datastore` plugin.

This commit relies on untagged support in the
atc0005/go-nagios project for handling
Performance Data. The state of that initial implementation
and the changes provided by this commit are still subject
to change based on testing/feedback.

refs GH-321
refs atc0005/go-nagios#81
refs atc0005/go-nagios@1db1acf6efc3e6907e7ab5bd6ba0437208f8df2a