Initial support for Nagios Performance Data output
Add initial support for emitting performance data metrics via the
`check_vmware_datastore` plugin. This support is provided using the
latest changes introduced in `v0.8.0` of the `atc0005/go-nagios`
package.

Two metrics are currently supported:

- `datastore_usage`
- `datastore_storage_remaining`

Further metrics for this plugin and initial support of performance
data for other plugins in this project are planned for future
releases of this project.

refs GH-321