Update README coverage for stderr output
Until recently, emitting plugin logging details to `stderr`
was seen as a safe approach. `stdout` is processed by Nagios
and used to generate notifications, web UI details, etc. Output
to `stderr` by default was intended to make troubleshooting
plugin operation easier for sysadmins.

After receiving a report from @Byolock, I learned that Icinga
Web 2 displays `stderr` output within the web UI mixed in with
`stdout`. This is undesirable.

This commit attempts to make both the original intent and
potential problems clearer along with providing a workaround
for the command definitions for plugins in this project:

--log-level "disabled"

refs GH-314
refs GH-323