Skip eval of powered off VMs without running Tools
VMs with a VMware Tools status of `toolsNotRunning` IS
problematic for powered on VMs, but not for powered off VMs as
this is the expected status. This commit updates the
`vsphere.FilterVMsWithToolsIssues()` function to ignore
VMs in this state.

Any powered off VMs with other VMware Tools statuses are still
evaluated as before.

NOTE: There are edge cases not currently handled by these
changes. See GH-366 for details.

refs GH-365
refs GH-366