Replace use of deprecated toolsStatus field
Update functions which rely on this field to determine VMware Tools
status to instead use the supported `toolsRunningStatus` and
`toolsVersionStatus2` fields.

Create helper function to consolidate evaluation logic and use it when
filtering the initial VirtualMachine collection and later for
determining the overall service state of the collection.

Details included in the `LongServiceOutput` have also been updated to
use the new field details and the stated threshold criteria have
similarly been updated to reflect updated logic.

refs GH-366