Initial refactor pass for `check_vmware_hs2ds2vms`
Overall, this is a collection of changes to simplify the `main()`
function in the `check_vmware_hs2ds2vms` plugin and make the logic
easier to follow. More work remains to be done in a future refactoring
effort, including reducing the swath of parameters provided to the
reporting functions.

CHANGES:

- Move `[]DatastoreWithCA` and `[]HostWithCA` retrieval to
  `vsphere.GetHostsWithCA()` and `vsphere.GetDatastoresWithCA()`
  functions

- Move `DatastoresPairingIssues` retrieval to new
  `vsphere.GetVMDatastorePairingIssues()` function

- Update doc comments, debug logging tweaks

FIXES:

- Replace wrong success debug message left behind from using VMware
  Tools plugin as a starter template

refs GH-394