Clarify intent of `timeout` flag
Update existing, add new doc comments to note that the
existing `timeout` flag and settings are specific to
plugin runtime/execution and not the initial connection
attempt to an ESXi server or vCenter instance.

Previously, the documentation noted plugin runtime/execution
while the doc comments noted connection timeout. This was
confusing and inconsistent.

refs GH-414