Add & use vsphere.AnnotateError() to wrap errors
- add `vsphere.ErrRuntimeTimeoutReached` error to indicate that an
  error has occurred due to a plugin's runtime exceeding a specified
  timeout value
- add `vsphere.AnnotateError()` func to *potentially* wrap passed
  errors with additional context
  - initial use case is annotating the `context.DeadlineExceeded`
    error type with `vsphere.ErrRuntimeTimeoutReached` error details
- update plugins currently emitting perf data to use
  `vsphere.AnnotateError()` func in a deferred anonymous func to wrap
  `(nagios.ExitState).LastError`

The end result is that these plugins now provide all of the error
output details as before, but also additional details if a
`context.DeadlineExceeded` error occurs:

- `check_vmware_datastore`
- `check_vmware_disk_consolidation`
- `check_vmware_host_cpu`
- `check_vmware_host_memory`
- `check_vmware_question`
- `check_vmware_snapshots_age`
- `check_vmware_snapshots_count`
- `check_vmware_snapshots_size`
- `check_vmware_tools`

Other plugins in the project slated for perf data work will have these
same changes applied as part of that work.

While this commit represents a functioning collection of changes, but
is likely not the final form; it is likely that some of this behavior
will be moved to the `atc0005/go-nagios` project at some point in the
future.

refs GH-463