Add initial draft of datastore performance plugin
This plugin monitors a specified datastore for performance metrics
which exceed latency thresholds.

This plugin accepts flags to:

- specify individual latency metric thresholds (e.g., read latency
  CRITICAL, read latency WARNING, write latency ...)
- specify percentile *sets*
  - multiple sets supported, each composed of a percentile and pairs
    of CRITICAL and WARNING threshold values

If not explicitly provided, default values are used.

Performance Data metrics are provided for each supported percentile:

- `90`
- `80`
- `70`
- `60`
- `50`

Performance Data metrics are omitted for a percentile when all
metrics in the percentile set are of value 0.

This plugin evaluates the active interval, but provides historical
interval metrics via `LongServiceOutput` unless explicitly requested
via a flag not to emit those details.

This plugin attempts to identify whether performance metrics
collection is not enabled for a datastore. If identified, an
`UNKNOWN` state is returned by the plugin as (reliable) evaluation
of a datastore's performance is not possible.

Doc updates have been applied, example usage has been added, including
a command definition "contrib" file illustrating how the plugin would
be referenced within a production Nagios configuration.

- refs GH-316
- refs GH-505