Rework `*_remaining` and `*_used` metrics
Updated plugins:

- check_vmware_datastore_space
- check_vmware_host_cpu
- check_vmware_host_memory
- check_vmware_rps_memory
- check_vmware_vcpus

Drop `Max` and `Min` qualifiers so that these metrics are based
solely on current values. The intent is to provide historical
details soley for planning purposes (without attaching a
specific severity to the values); the `*_usage` metrics remain
the intended metric for determining severity.

Update CHANGELOG to note this is a breaking change.

refs GH-480