Update logger init to better note excluded root rp
Setup logger initially as with other plugins, then state
the intent to explicitly exclude the default/parent/root
resource pool. After successfully doing so, redefine the
logger so that the newly added explicit exclusion is
emitted for log messages going forward.

The modifications are done in an attempt to better clarify
which exclusions the sysadmin specified and which the
plugin explicitly set (currently just the root pool).

Further refactoring may be needed to better handle setting
and noting this choice.