Adjust make all recipe job timeout
Increase from 40 minutes to 55 minutes to handle recent timeouts.

The cause of the timeouts was assumed to be related to an upgrade
of the Go toolchain from 1.17.13 to 1.19.1, but sufficient
testing has not been performed to conclusively prove this is
the case. This timeout adjustment is considered as a hotfix
until additional testing can be performed.

refs GH-724